Fix one occassion of Werror=address
With GCC 12 gcc started to detect if comparisons to addresses are always
true. This could indicate a bug, as seen in the Linux kernel. For
instance, a simple overlooked ampersand could turn a useful comparison
to a tautological default, e.g. in `&my_struct->my_value`.

In this particular case it seems as if the comparison was superfluous.